requirements.txt
torch
numpy
panda
sklearn
matplotlib
tqdm